% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CALIBRATION.R
\name{CALIBRATION}
\alias{CALIBRATION}
\title{Calibration
Takes in RFUs. Output correponding RFUs after Calibration normalization.}
\usage{
CALIBRATION(RawM)
}
\arguments{
\item{RawM}{the input RFUs data frame from SomaLogic Adat file.}
}
\value{
MySoma the output RFUs data frame after the Calibration normalization.
}
\description{
Calibration
Takes in RFUs. Output correponding RFUs after Calibration normalization.
}
