% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMySoma.R
\name{getMySoma}
\alias{getMySoma}
\title{Combine platewise normalised data frame
Takes in a list with element of normalised data frame per plate. Output a whole data frame combining the element of Platelist.}
\usage{
getMySoma(Platelist)
}
\arguments{
\item{Platelist}{the input RFUs data frame from SomaLogic Adat file.}
}
\value{
MySoma the whole RFUs data frame containing all the plates
}
\description{
Combine platewise normalised data frame
Takes in a list with element of normalised data frame per plate. Output a whole data frame combining the element of Platelist.
}
