% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HybNorm.R
\name{HYBNORM}
\alias{HYBNORM}
\title{Hybridizatioin normalisation
Takes in RFUs. Output correponding RFUs after Hybridization normalization.}
\usage{
HYBNORM(RawM)
}
\arguments{
\item{RawM}{the input RFUs data frame from SomaLogic Adat file.}
}
\value{
MySoma the output RFUs data frame after the hybridization normalization.
}
\description{
Hybridizatioin normalisation
Takes in RFUs. Output correponding RFUs after Hybridization normalization.
}
