% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initQCnorm.R
\name{initQCnorm}
\alias{initQCnorm}
\title{initialise required metadata for normalisation methods
Takes in raw adat RFUs files and the other RFUs files containing external reference. Output RawM the raw RFUs data frame for normalisation.}
\usage{
initQCnorm(inputfile1, inputfile2)
}
\arguments{
\item{inputfile1}{SomaLogic Adat file, which contains the raw RFUs}

\item{inputfile2}{SomaLogic Adat file, which contains metadata.}
}
\value{
RawM the raw RFUs data frame for normalisation.
}
\description{
initialise required metadata for normalisation methods
Takes in raw adat RFUs files and the other RFUs files containing external reference. Output RawM the raw RFUs data frame for normalisation.
}
