% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MIDNORM.R
\name{MIDNORM}
\alias{MIDNORM}
\title{Median normalisation
Takes in RFUs. Output correponding RFUs after Median normalization.}
\usage{
MIDNORM(RawM)
}
\arguments{
\item{RawM}{the input RFUs data frame from SomaLogic Adat file.}
}
\value{
MySoma the output RFUs data frame after the Median normalisation.
}
\description{
Median normalisation
Takes in RFUs. Output correponding RFUs after Median normalization.
}
