% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MIDNORM.R
\name{MIDNORMcali}
\alias{MIDNORMcali}
\title{Median normalisation on calibrators
Takes in RFUs. Output correponding RFUs after Median normalization.}
\usage{
MIDNORMcali(RawM)
}
\arguments{
\item{RawM}{the input RFUs data frame from SomaLogic Adat file.}
}
\value{
MySoma the output RFUs data frame after the Median normalisation on calibrators.
}
\description{
Median normalisation on calibrators
Takes in RFUs. Output correponding RFUs after Median normalization.
}
