% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MIDNORM.R
\name{MIDNORMsamp}
\alias{MIDNORMsamp}
\title{Median normalisation on samples
Takes in RFUs. Output correponding RFUs after Median normalization.}
\usage{
MIDNORMsamp(RawM)
}
\arguments{
\item{RawM}{the input RFUs data frame from SomaLogic Adat file.}
}
\value{
MySoma the output RFUs data frame after the Median normalisation on samples
}
\description{
Median normalisation on samples
Takes in RFUs. Output correponding RFUs after Median normalization.
}
