% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PLATESCALE.R
\name{PLATESCALE}
\alias{PLATESCALE}
\title{Plate scaling
Takes in RFUs. Output correponding RFUs after plate scaling normalization.}
\usage{
PLATESCALE(RawM)
}
\arguments{
\item{RawM}{the input RFUs data frame from SomaLogic Adat file.}
}
\value{
MySoma the output RFUs data frame after the plate scaling normalization.
}
\description{
Plate scaling
Takes in RFUs. Output correponding RFUs after plate scaling normalization.
}
