% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UserNorm.R
\name{UserNorm}
\alias{UserNorm}
\title{Visible to users
Takes in the raw RFUs data frame from SomaLogic adat file, and also user defined normalisation workflow.}
\usage{
UserNorm(Funlist, RawM)
}
\arguments{
\item{Funlist}{user define a list: select from "HYBNORM,PLATESCALE,MIDNORM/MIDNORMcali/MIDNORMsamp,CALIBRATION.}

\item{RawM}{the input RFUs data frame from SomaLogic Adat file.}
}
\value{
MySoma the normalised RFUs after user defined normalisation procedures.
}
\description{
Visible to users
Takes in the raw RFUs data frame from SomaLogic adat file, and also user defined normalisation workflow.
}
